@auth with owner auth. New tests. Adding libs to .gitignore

+++ packages/amplify-graphql-mapping-template/lib/dynamodb.js
@@ -65,11 +65,12 @@ var DynamoDBMappingTemplate = /** @class */ (function () {
      * @param key A list of strings pointing to the key value locations. E.G. ctx.args.x (note no $)
      */
     DynamoDBMappingTemplate.deleteItem = function (_a) {
-        var key = _a.key;
+        var key = _a.key, condition = _a.condition;
         return ast_1.obj({
             version: ast_1.str('2017-02-28'),
             operation: ast_1.str('DeleteItem'),
-            key: key
+            key: key,
+            condition: condition,
         });
     };
     /**
@@ -79,20 +80,23 @@ var DynamoDBMappingTemplate = /** @class */ (function () {
     DynamoDBMappingTemplate.updateItem = function (_a) {
         var key = _a.key, condition = _a.condition;
         var keyNames = key.attributes.map(function (attr) { return attr[0]; });
+        // Auto timestamp
+        // qref('$input.put("updatedAt", "$util.time.nowISO8601()")'),
         return ast_1.compoundExpression([
+            ast_1.iff(ast_1.raw('!$input'), ast_1.set(ast_1.ref('input'), ast_1.ref('util.map.copyAndRemoveAllKeys($context.args.input, [])'))),
             ast_1.set(ast_1.ref('expNames'), ast_1.obj({})),
             ast_1.set(ast_1.ref('expValues'), ast_1.obj({})),
             ast_1.set(ast_1.ref('expSet'), ast_1.obj({})),
             ast_1.set(ast_1.ref('expAdd'), ast_1.obj({})),
             ast_1.set(ast_1.ref('expRemove'), ast_1.list([])),
-            ast_1.forEach(ast_1.ref('entry'), ast_1.ref("util.map.copyAndRemoveAllKeys($ctx.args.input, [" + keyNames.map(function (k) { return "\"" + k + "\""; }).join(', ') + "]).entrySet()"), [
+            ast_1.forEach(ast_1.ref('entry'), ast_1.ref("util.map.copyAndRemoveAllKeys($input, [" + keyNames.map(function (k) { return "\"" + k + "\""; }).join(', ') + "]).entrySet()"), [
                 ast_1.ifElse(ast_1.ref('util.isNull($entry.value)'), ast_1.compoundExpression([
-                    ast_1.set(ast_1.ref('discard'), ast_1.ref('expRemove.add("#${entry.key}")')),
-                    ast_1.qref('$expNames.put("#${entry.key}", "${entry.key}")')
+                    ast_1.set(ast_1.ref('discard'), ast_1.ref('expRemove.add("#$entry.key")')),
+                    ast_1.qref('$expNames.put("#$entry.key", "$entry.key")')
                 ]), ast_1.compoundExpression([
-                    ast_1.qref('$expSet.put("#${entry.key}", ":${entry.key}")'),
-                    ast_1.qref('$expNames.put("#${entry.key}", "${entry.key}")'),
-                    ast_1.qref('$expValues.put(":${entry.key}", $util.dynamodb.toDynamoDB($entry.value))')
+                    ast_1.qref('$expSet.put("#$entry.key", ":$entry.key")'),
+                    ast_1.qref('$expNames.put("#$entry.key", "$entry.key")'),
+                    ast_1.qref('$expValues.put(":$entry.key", $util.dynamodb.toDynamoDB($entry.value))')
                 ]))
             ]),
             ast_1.set(ast_1.ref('expression'), ast_1.str('')),
@@ -104,14 +108,14 @@ var DynamoDBMappingTemplate = /** @class */ (function () {
                 ])
             ])),
             ast_1.iff(ast_1.raw('!$expAdd.isEmpty()'), ast_1.compoundExpression([
-                ast_1.set(ast_1.ref('expression'), ast_1.str('${expression} ADD')),
+                ast_1.set(ast_1.ref('expression'), ast_1.str('$expression ADD')),
                 ast_1.forEach(ast_1.ref('entry'), ast_1.ref('expAdd.entrySet()'), [
                     ast_1.set(ast_1.ref('expression'), ast_1.str('$expression $entry.key $entry.value')),
                     ast_1.iff(ast_1.ref('foreach.hasNext()'), ast_1.set(ast_1.ref('expression'), ast_1.str('$expression,')))
                 ])
             ])),
             ast_1.iff(ast_1.raw('!$expRemove.isEmpty()'), ast_1.compoundExpression([
-                ast_1.set(ast_1.ref('expression'), ast_1.str('${expression} REMOVE')),
+                ast_1.set(ast_1.ref('expression'), ast_1.str('$expression REMOVE')),
                 ast_1.forEach(ast_1.ref('entry'), ast_1.ref('expRemove'), [
                     ast_1.set(ast_1.ref('expression'), ast_1.str('$expression $entry')),
                     ast_1.iff(ast_1.ref('foreach.hasNext()'), ast_1.set(ast_1.ref('expression'), ast_1.str('$expression,')))
@@ -119,8 +123,8 @@ var DynamoDBMappingTemplate = /** @class */ (function () {
             ])),
             ast_1.set(ast_1.ref('update'), ast_1.obj({})),
             ast_1.qref('$update.put("expression", "$expression")'),
-            ast_1.iff(ast_1.raw('!${expNames.isEmpty()}'), ast_1.qref('$update.put("expressionNames", $expNames)')),
-            ast_1.iff(ast_1.raw('!${expValues.isEmpty()}'), ast_1.qref('$update.put("expressionValues", $expValues)')),
+            ast_1.iff(ast_1.raw('!$expNames.isEmpty()'), ast_1.qref('$update.put("expressionNames", $expNames)')),
+            ast_1.iff(ast_1.raw('!$expValues.isEmpty()'), ast_1.qref('$update.put("expressionValues", $expValues)')),
             ast_1.obj({
                 version: ast_1.str('2017-02-28'),
                 operation: ast_1.str('UpdateItem'),
